Move src.silx.utils.cache to src.gui.plot.items._cache
