+++ src/silx/gui/plot/items/_cache.py
